+++ eltainer.el
@@ -36,6 +36,7 @@
     ("Kubernetes" .
      (("P" "Pulse"         k8s-pulse)
       ("E" "Events"        k8s-events)
+      ("Z" "Scan (sanity)" k8s-scan)
       ("k" "Pods"          k8s-pods)
       ("d" "Deployments"   k8s-deployments)
       ("s" "Services"      k8s-services)